Release: WHO Adapt Sprint 104 (#104)

+++ frontend/src/modules/editor/contentObject/views/editorPageComponentView.js
@@ -41,17 +41,26 @@ define(function(require){
       }, this));
     },
 
-    deleteComponentPrompt: function(event) {
-      event && event.preventDefault();
+    deleteComponentPrompt: async function(event) {
+      const currentUserRole = await Origin.getCurrentUserRole();
+      // if the user is authenticated should not be able to delete components, show alert permission denied
+      if (currentUserRole === 'Authenticated User') {
+        Origin.Notify.alert({
+          type: 'error',
+          text: 'You do not have permission to edit or delete the courses'
+        });
+      } else {
+        event && event.preventDefault();
 
-      Origin.Notify.confirm({
-        type: 'warning',
-        title: Origin.l10n.t('app.deletecomponent'),
-        text: Origin.l10n.t('app.confirmdeletecomponent') + '<br />' + '<br />' + Origin.l10n.t('app.confirmdeletecomponentwarning'),
-        callback: _.bind(function(confirmed) {
-          if(confirmed) this.deleteComponent();
-        }, this)
-      });
+        Origin.Notify.confirm({
+          type: 'warning',
+          title: Origin.l10n.t('app.deletecomponent'),
+          text: Origin.l10n.t('app.confirmdeletecomponent') + '<br />' + '<br />' + Origin.l10n.t('app.confirmdeletecomponentwarning'),
+          callback: _.bind(function(confirmed) {
+            if(confirmed) this.deleteComponent();
+          }, this)
+        });
+      }
     },
 
     deleteComponent: function() {
